Изменяет название модулей с играми
