+++ js/game-one.js
@@ -1,6 +1,6 @@
 import {getElementFromTemplate, changeScreen} from './services';
 import backToGreeting from './back';
-import gameThree from './game-3';
+import gameTwo from './game-two';
 
 const TEMPLATE = `
   <header class="header">
@@ -15,22 +15,33 @@ const TEMPLATE = `
     </button>
     <div class="game__timer">NN</div>
     <div class="game__lives">
-      <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="31" height="27">
+      <img src="img/heart__empty.svg" class="game__heart" alt=" Missed Life" width="31" height="27">
       <img src="img/heart__full.svg" class="game__heart" alt="Life" width="31" height="27">
       <img src="img/heart__full.svg" class="game__heart" alt="Life" width="31" height="27">
     </div>
   </header>
   <section class="game">
-    <p class="game__task">Угадай, фото или рисунок?</p>
-    <form class="game__content  game__content--wide">
+    <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+    <form class="game__content">
       <div class="game__option">
-        <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-        <label class="game__answer  game__answer--photo">
+        <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+        <label class="game__answer game__answer--photo">
           <input class="visually-hidden" name="question1" type="radio" value="photo">
           <span>Фото</span>
         </label>
-        <label class="game__answer  game__answer--paint">
+        <label class="game__answer game__answer--paint">
           <input class="visually-hidden" name="question1" type="radio" value="paint">
+          <span>Рисунок</span>
+        </label>
+      </div>
+      <div class="game__option">
+        <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+        <label class="game__answer  game__answer--photo">
+          <input class="visually-hidden" name="question2" type="radio" value="photo">
+          <span>Фото</span>
+        </label>
+        <label class="game__answer  game__answer--paint">
+          <input class="visually-hidden" name="question2" type="radio" value="paint">
           <span>Рисунок</span>
         </label>
       </div>
@@ -40,26 +51,29 @@ const TEMPLATE = `
       <li class="stats__result stats__result--slow"></li>
       <li class="stats__result stats__result--fast"></li>
       <li class="stats__result stats__result--correct"></li>
-      <li class="stats__result stats__result--wrong"></li>
       <li class="stats__result stats__result--unknown"></li>
-      <li class="stats__result stats__result--slow"></li>
       <li class="stats__result stats__result--unknown"></li>
-      <li class="stats__result stats__result--fast"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--unknown"></li>
       <li class="stats__result stats__result--unknown"></li>
     </ul>
   </section>
 `;
 
-const gameTwo = getElementFromTemplate(TEMPLATE);
-const gameContent = gameTwo.querySelector(`.game__content`);
-const back = gameTwo.querySelector(`.back`);
+const gameOne = getElementFromTemplate(TEMPLATE);
+const gameContent = gameOne.querySelector(`.game__content`);
+const back = gameOne.querySelector(`.back`);
 
-const answerHandler = () => {
-  changeScreen(gameThree);
+const answersHandler = () => {
+  const answers = gameOne.querySelectorAll(`input[type=radio]:checked`);
+  if (answers.length === 2) {
+    changeScreen(gameTwo);
+  }
 };
 
-gameContent.addEventListener(`input`, answerHandler);
+gameContent.addEventListener(`input`, answersHandler);
 backToGreeting(back);
 
 
-export default gameTwo;
+export default gameOne;
